moved top feed to summary
